total rewrite of the clan promotion system.  god help all the clans tonight.
git-svn-id: svn://192.168.1.10/public/CoffeeMud@17725 0d6f1817-ed0e-0410-87c9-987e46238f29

+++ com/planet_ink/coffee_mud/Common/DefaultClanPosition.java
@@ -57,7 +57,7 @@ public class DefaultClanPosition implements ClanPosition
 	/** the plural name of the position within this government */
 	protected String	pluralName;
 	/** the maximum number of members that can hold this position */
-	protected int		max;
+	protected double	max;
 	/** the internal zapper mask for internal requirements to this position */
 	protected String	innerMaskStr;
 	/** the internal zapper mask for internal requirements to this position */
@@ -166,13 +166,13 @@ public class DefaultClanPosition implements ClanPosition
 	}
 
 	@Override
-	public int getMax()
+	public double getMax()
 	{
 		return max;
 	}
 
 	@Override
-	public void setMax(final int max)
+	public void setMax(final double max)
 	{
 		this.max = max;
 	}
@@ -257,7 +257,7 @@ public class DefaultClanPosition implements ClanPosition
 		case RANK:
 			return Integer.toString(rank);
 		case MAX:
-			return Integer.toString(max);
+			return Double.toString(max);
 		case PLURALNAME:
 			return pluralName;
 		case INNERMASK:
